% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{boot_change_point}
\alias{boot_change_point}
\title{Bootstrap Estimation of Changepoint and simulated miss visits}
\usage{
boot_change_point(
  sim_data,
  n_sim_trials = 100L,
  sim_algorithm = "simple",
  sim_ctrl = NULL,
  eval_criteria = "AIC",
  week_period = FALSE,
  num_cores = NULL,
  no_bootstrapping = FALSE,
  auto_reg = FALSE
)
}
\arguments{
\item{sim_data}{a dataset containing the time_map, miss bins and other parameters used for the simulation.
This dataset should be created using the `prep_sim_data()` function}

\item{n_sim_trials}{Number of trials to run in simulation of miss visits or miss patients}

\item{sim_algorithm}{a string defining the algorithm to use. Option include "simple", 
"simple_correlated" and "general"}

\item{sim_ctrl}{the sim_ctrl function that defines additional parameter values passed to the 
simulation. If no value is supplied (i.e., sim_ctrl = NULL) then default 
values for the sim_ctrl() function will be used}

\item{num_cores}{The number of worker cores to use. If not specified will detect cores and use 1 less than the number of cores}

\item{no_bootstrapping}{Specifies whether you want to run the simulations without bootstrapping the original dataset}
}
\description{
Bootstrap Estimation of Changepoint and simulated miss visits
}
