% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCUP_timemap_funs.R
\name{build_HCUP_time_map}
\alias{build_HCUP_time_map}
\title{Build Delay Time Map with ASCII HCUP Data}
\usage{
build_HCUP_time_map(path, state, year, index_codes, ssd_list, n_collect)
}
\arguments{
\item{path}{(Character) Path to ASCII files}

\item{state}{(Character) Two-letter abbervation for the desired state}

\item{year}{(Character) Typically, a four digit expression of desired year}

\item{index_codes}{(list) Typically, a list length 2 for ICD-9-CM and ICD-10-CM codes for the condition of interest}

\item{ssd_list}{(tibble) Typically, a tibble containing the SSD names and their respective ICD-9-CM and ICD-10-CM codes}

\item{n_collect}{(Numeric) Number of rows to read in. Defaults to infinity.}
}
\value{
a tibble for the time map of interest
}
\description{
Build Delay Time Map with ASCII HCUP Data
}
