% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_output_functions.R
\name{compute_miss_num}
\alias{compute_miss_num}
\title{Compute number of missed visits per patient}
\usage{
compute_miss_num(miss_draw_data)
}
\description{
Compute number of missed visits per patient
}
