% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_output_functions.R
\name{compute_sim_trial_duration_table}
\alias{compute_sim_trial_duration_table}
\title{Compute simulation trial duration table}
\usage{
compute_sim_trial_duration_table(sim_miss_num_data, upper_bound)
}
\description{
Compute simulation trial duration table
}
