% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_output_functions.R
\name{compute_sim_trial_n_visit_table}
\alias{compute_sim_trial_n_visit_table}
\title{Compute simulation trial visit table}
\usage{
compute_sim_trial_n_visit_table(sim_miss_num_data, total_patients)
}
\description{
Compute simulation trial visit table
}
