% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCUP_timemap_funs.R
\name{find_data_version}
\alias{find_data_version}
\title{Determine Data Version Number}
\usage{
find_data_version(path, file_type, data_source, state, year)
}
\arguments{
\item{path}{(Character) Path to CORE ASCII file}

\item{file_type}{(Character) Which file to parse (e.g., core, ahal, severity)}

\item{data_source}{(Character) The data set source, either SEDD or SID}

\item{state}{(Character) Two-letter abbervation for the desired state}

\item{year}{(Character) Typically, a four digit expression of desired year
but may be 2015q1q3 or 2015q4 to accomdate the switch from ICD-9 to ICD-10
on 2015/10/01}
}
\value{
a vector of length 1 with the version number of the data located at
\code{path} based on the observed line-length and the expected line length
for the different versions of data provided for \code{data_source},
\code{state} and \code{year}. A version of \code{0} denotes the current
version while values of \code{-1}, \code{-2} and so on denote versions
1, 2 and so on.
}
\description{
Determine Data Version Number
}
