% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCUP_timemap_funs.R
\name{find_variable_locations}
\alias{find_variable_locations}
\title{Find Variables and Locations by Data Souce, State and Year}
\usage{
find_variable_locations(data_source, state, year)
}
\arguments{
\item{data_source}{character The data set source, either SEDD or SID}

\item{state}{character Two-letter abbervation for the desired state}

\item{year}{character Typically, a four digit expression of desired year
but may be 2015q1q3 or 2015q4 to accomdate the switch from ICD-9 to ICD-10
on 2015/10/01}
}
\value{
a tibble with columns for variable name, encoding type, start and
end columns referring to the ASCII file for data source, state and year.
}
\description{
Find Variables and Locations by Data Souce, State and Year
}
