% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCUP_timemap_funs.R
\name{get_SID}
\alias{get_SID}
\title{Get SID helper functions used in build_HCUP_time_map}
\usage{
get_SID(data, index_codes, ssd_list)
}
\arguments{
\item{data}{(tibble) Typically, a tibble containing the pasrsed SID data from the ASCII file}

\item{index_codes}{(list) Typically, a list length 2 for ICD-9-CM and ICD-10-CM codes for the condition of interest}

\item{ssd_list}{(tibble) Typically, a tibble containing the SSD names and their respective ICD-9-CM and ICD-10-CM codes}
}
\value{
a tibble for the time map of interest
}
\description{
Get SID helper functions used in build_HCUP_time_map
}
