% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cp_methods_functions.R
\name{MSE}
\alias{MSE}
\title{Mean squared error}
\usage{
MSE(predicted, observed)
}
\arguments{
\item{predicted}{Vector of predicted values}

\item{observed}{Vector of observed values}
}
\value{
A scalar value
}
\description{
Mean squared error
}
