% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cp_methods_functions.R
\name{MSLE}
\alias{MSLE}
\title{Mean squared log error}
\usage{
MSLE(predicted, observed)
}
\arguments{
\item{predicted}{Vector of predicted values}

\item{observed}{Vector of observed values}
}
\value{
A scalar value
}
\description{
Mean squared log error
}
