% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCUP_timemap_funs.R
\name{parse_data}
\alias{parse_data}
\title{Parse ASCII Files}
\usage{
parse_data(path, file_type, data_source, state, year, n_max = Inf)
}
\arguments{
\item{path}{(Character) Path to ASCII file}

\item{file_type}{(Character) Which file to parse (e.g., core, ahal, severity)}

\item{data_source}{(Character) The data set source, either SEDD or SID}

\item{state}{(Character) Two-letter abbervation for the desired state}

\item{year}{(Character) Typically, a four digit expression of desired year
but may be 2015q1q3 or 2015q4 to accomdate the switch from ICD-9 to ICD-10
on 2015/10/01}

\item{n_max}{(Numeric) Number of rows to read in. Defaults to infinity.
Useful for testing.}
}
\value{
A tibble of the parsed data from the data located at \code{path}
using the parameters detailed by \code{file_type}, \code{data_source},
\code{state}, \code{year} and the version guessed by
\code{find_data_version()}.
}
\description{
Parse ASCII Files
}
