% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\name{prep_cp_data}
\alias{prep_cp_data}
\title{Count prior visits assocatied with a given condition for preparation in finding a changepoint}
\usage{
prep_cp_data(time_map_data, event_name, start_day = 1L, by_days = 1L)
}
\arguments{
\item{time_map_data}{A timemap with indicators for specific events (i.e.
diagnoses during a given visit)}

\item{event_name}{The variable name for the even indicator}

\item{start_day}{When to start counting prior to the index}

\item{by_days}{The number of days in each bin to count by}
}
\description{
Count prior visits assocatied with a given condition for preparation in finding a changepoint
}
