% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cp_methods_functions.R
\name{RMSE}
\alias{RMSE}
\title{Root mean squared error}
\usage{
RMSE(predicted, observed)
}
\arguments{
\item{predicted}{Vector of predicted values}

\item{observed}{Vector of observed values}
}
\value{
A scalar value
}
\description{
Root mean squared error
}
