% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cp_methods_functions.R
\name{RMSLE}
\alias{RMSLE}
\title{Root mean squared log error}
\usage{
RMSLE(predicted, observed)
}
\arguments{
\item{predicted}{Vector of predicted values}

\item{observed}{Vector of observed values}
}
\value{
A scalar value
}
\description{
Root mean squared log error
}
