% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cp_methods_functions.R
\name{run_compare_method}
\alias{run_compare_method}
\title{Compare change point methods}
\usage{
run_compare_method(sim_data, method1, method2, loss_fun = "MSE")
}
\arguments{
\item{sim_data}{A dataset containing the time_map, miss bins and other parameters used for the simulation.
This dataset should be created using the `prep_sim_data()` function}

\item{method1}{First change point method to compare}

\item{method2}{Second change point method to compare}

\item{loss_fun}{The loss function used in the comparison (mean squared error (MSE), root mean squared error (RMSE),
mean absolute error (MAE), mean squared log error (MSLE), or root mean squared log error (RMSLE)}
}
\value{
A one row tibble which 3 columns: 1) the loss function used, 2) the value for method 1, and 3) the value for method 2
}
\description{
Compare change point methods
}
