% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_helper_functions.R
\name{run_sim_draw_general}
\alias{run_sim_draw_general}
\title{Draw Missed Visits Using the generalized algorithm}
\usage{
run_sim_draw_general(sim_data, sim_ctrl)
}
\arguments{
\item{sim_ctrl}{the control function specifying the simulation parameters. This simulation uses
the parameters...}
}
\description{
Draw Missed Visits Using the generalized algorithm
}
