% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_helper_functions.R
\name{run_sim_draw_simple}
\alias{run_sim_draw_simple}
\title{Draw Missed Visits Using the simple algorithm}
\usage{
run_sim_draw_simple(sim_data)
}
\description{
Draw Missed Visits Using the simple algorithm
}
