% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{run_sim_miss_visits}
\alias{run_sim_miss_visits}
\title{Run multiple simulations of missed visits}
\usage{
run_sim_miss_visits(
  sim_data,
  trials = 50,
  no_bootstrapping = FALSE,
  sim_algorithm = "simple",
  sim_ctrl = NULL,
  num_cores = NULL
)
}
\arguments{
\item{sim_data}{a dataset containing the time_map, miss bins and other parameters used for the simulation.
This dataset should be created using the `prep_sim_data()` function}

\item{trials}{number of simulationed trials to run (default is 50)}

\item{no_bootstrapping}{Specifies whether you want to run the simulations without bootstrapping the original dataset}

\item{sim_algorithm}{a string defining the algorithm to use. Option include "simple", 
"simple_correlated" and "general"}

\item{sim_ctrl}{the sim_ctrl function that defines additional parameter values passed to the 
simulation. If no value is supplied (i.e., sim_ctrl = NULL) then default 
values for the sim_ctrl() function will be used}

\item{num_cores}{The number of worker cores to use. If not specified will determined the number of cores based on the which ever
is the smallest value between number of trials or detected number of cores - 1}
}
\description{
Run multiple simulations of missed visits
}
