% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_functions.R
\name{set_change_point}
\alias{set_change_point}
\title{set_change_point}
\usage{
set_change_point(
  data,
  var_name = "n_miss_visits",
  method = NULL,
  compare_all_methods = FALSE,
  return_miss_only = FALSE,
  week_period = FALSE,
  specify_cp = NULL
)
}
\arguments{
\item{data}{A dataset of visit counts, output by count_prior_events_truven}

\item{var_name}{The name of the count variable to find the change-point for}

\item{method}{The method used to find changepoint. Options include "linear", "quadratic", "cubic", "quartic",
"quintic", "sextic" , or "exponential"}

\item{compare_all_methods}{A logical to compare all available. If TRUE, will return a tibble of methods by evaluation criteria}

\item{return_miss_only}{Logical argument to only return the tibbles of miss visit counts}

\item{week_period}{Logical to incorporate a "day of the week" effect into the momdel
Note this is only sensible for one-day period aggregation}

\item{specify_cp}{A postive integer value for the specific change point you want to use. The value represents
the days before the index on which you you want to specify the change point. (e.g. 100 would be 100 days before the index)}

\item{eval_criteria}{The evaluation criteria used to compare models}
}
\value{
A list containing tibbles of information about missed visits
}
\description{
Set the change point and identify the optimal method to model the data prior to the change point
}
\examples{

results <- final_time_map \%>\%
count_prior_events_truven(event_name = "any_ssd", start_day = 1, by_days = 1) \%>\%
set_change_point(var_name = "n_miss_visits", method = "cubic", specify_cp = 21)

}
