% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_functions.R
\name{set_cp_poly}
\alias{set_cp_poly}
\title{set_cp_poly}
\usage{
set_cp_poly(
  data,
  var_name = "n_miss_visits",
  return_miss_only = FALSE,
  return_eval_only = FALSE,
  week_period = FALSE,
  specify_cp = NULL,
  poly_order = "linear",
  use_orthogonal_poly = FALSE
)
}
\arguments{
\item{data}{A dataframe output by count_prior_events_truven}

\item{var_name}{A character string of outcome for which to apply analysis}

\item{return_miss_only}{Logical to only return miss information}

\item{return_eval_only}{Logical to only return model evaluation criteria info}

\item{week_period}{Logical to incorporate a "day of the week" effect into model. Note this is only sensible for one-day period aggregation.}

\item{specify_cp}{A postive integer value for the specific change point you want to use. The value represents
the days before the index on which you you want to specify the change point. (e.g. 100 would be 100 days before the index)}

\item{poly_order}{The polynomial order you want to apply (i.e. 'linear', 'quadratic', 'cubic', 'quartic', 'quintic', or 'sextic')}

\item{use_orthogonal_poly}{A logical to use raw or orthogonal polynomials. TRUE uses orthogonal polynomials. not raw}
}
\value{
A list containing miss information, changepoint information, predictions,
the model itself, and a plot of the middle finger curve and model
}
\description{
Set a change point and fit a model with a specified polynomial order on data prior to the change point to obtian expected SSD visits/calculate misses
}
