% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_sim_functions.R
\name{sim_miss_visits}
\alias{sim_miss_visits}
\title{Simulated miss visits from an estimated set of miss bins}
\usage{
sim_miss_visits(sim_data, sim_algorithm = "simple", sim_ctrl = NULL)
}
\arguments{
\item{sim_data}{a dataset containing the time_map, miss bins and other parameters used for the simulation.
This dataset should be created using the `prep_sim_data()` function}
}
\description{
Simulated miss visits from an estimated set of miss bins
}
