% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_helper_functions.R
\name{simple_weight_sum}
\alias{simple_weight_sum}
\title{simple weight function function for algorithm 3}
\usage{
simple_weight_sum(num_distinct_visits, num_distinct_ssds, num_prior_draws)
}
\arguments{
\item{num_distinct_visits}{number of distinct visits up to the time-point}

\item{num_distinct_ssds}{number of distinct ssds up to the time-point}

\item{num_prior_draws}{number of times the individual has been previously drawn}
}
\description{
simple weight function function for algorithm 3
}
