% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_helper_functions.R
\name{run_sim_draw_simple_cor}
\alias{run_sim_draw_simple_cor}
\title{Draw Missed Visits Using the simple correlated draw algorithm}
\usage{
run_sim_draw_simple_cor(sim_data, sim_ctrl)
}
\arguments{
\item{sim_ctrl}{the control function specifying the simulation parameters. This simulation uses
the parameter "alpha"}
}
\description{
Draw Missed Visits Using the simple correlated draw algorithm
}
