% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_sim_functions.R
\name{sim_ctrl}
\alias{sim_ctrl}
\title{Control function for specifying simulation parameters}
\usage{
sim_ctrl(alpha = 0.5)
}
\arguments{
\item{alpha}{draw parameter for simple_correlated simulation}
}
\description{
Control function for specifying simulation parameters
}
